+++ Modules/DataModel/Sources/PocketCastsDataModel/Public/Model/Episode.swift
@@ -209,11 +209,19 @@ public class Episode: NSObject, BaseEpisode {
         }
 
         public let transcripts: [Transcript]
+        public let pocketCastsTranscripts: [Transcript]?
 
         public struct Transcript: Decodable {
             public let url: String
             public let type: String
             public let language: String?
+        }
+
+        public func hasGeneratedTranscripts() -> Bool {
+            guard let pocketCastsTranscripts else {
+                return false
+            }
+            return !pocketCastsTranscripts.isEmpty
         }
     }
 }