+++ Modules/DataModel/Tests/PocketCastsDataModelTests/AutoAddCandidatesDataManagerTests.swift
@@ -14,7 +14,7 @@ final class AutoAddCandidatesDataManagerTests: XCTestCase {
 
     private func setupDatabase() throws -> DataManager {
         let dbQueue = try XCTUnwrap(FMDatabaseQueue.newTestDatabase())
-        return DataManager(dbQueue: dbQueue)
+        return DataManager(dbQueue: FMDBQueue(fmdbQueue: dbQueue))
     }
 
     /// Tests new query and autoAddToUpNext property for UpNext candidates