+++ Modules/DataModel/Tests/PocketCastsDataModelTests/PodcastDataManagerTests.swift
@@ -6,7 +6,7 @@ import FMDB
 final class PodcastDataManagerTests: XCTestCase {
     private func setupDatabase() throws -> DataManager {
         let dbQueue = try XCTUnwrap(FMDatabaseQueue.newTestDatabase())
-        return DataManager(dbQueue: dbQueue)
+        return DataManager(dbQueue: FMDBQueue(fmdbQueue: dbQueue))
     }
 
     private func setupDataManager() throws -> DataManager {