+++ podcasts/FoldersCoordinator.swift
@@ -3,36 +3,78 @@ import SwiftUI
 import PocketCastsDataModel
 import PocketCastsServer
 import PocketCastsUtils
+import Combine
 
 class FoldersCoordinator: NSObject {
 
+    enum UpsellFlow {
+        case none
+        case cta
+        case userInitiated
+    }
+
+    private var currentUpsellFlow: UpsellFlow = .none
+    private weak var currentVC: UIViewController? = nil
+
+    private var currentSource: AnalyticsSource = .unknown
+
+    private let startingTime: Date = Date.now
+
     private let navigationManager: NavigationManager
     private let dataManager: DataManager
+    private let suggestedFoldersModel: SuggestedFoldersModel
+
+    private enum Constants {
+        static let minimumNumberOfPodcasts: Int = 7
+        static let intervalBetweenUpsell: TimeInterval = 7.days
+        static let maxUpsellDisplays: Int = 2
+        static let intervalAfterStartup: TimeInterval = 10.seconds
+    }
 
     init(navigationManager: NavigationManager = .sharedManager, dataManager: DataManager = .sharedManager) {
         self.navigationManager = navigationManager
         self.dataManager = dataManager
+        self.suggestedFoldersModel = SuggestedFoldersModel()
+        super.init()
+        Task {
+            await suggestedFoldersModel.load()
+        }
     }
 
     func startFolderCreationFlow(from vc: UIViewController) {
-        if FeatureFlag.suggestedFolders.enabled {
-            newSuggestedFolderCreationFlow(from: vc)
+        if FeatureFlag.suggestedFolders.enabled, suggestedFoldersModel.loadingState == .loaded {
+            suggestedFolderCreationFlow(from: vc, source: .podcastsList)
         } else {
-            oldFolderCreationFlow(from: vc)
+            manualFolderCreationFlow(from: vc)
         }
         AnalyticsHelper.folderCreated()
         Analytics.track(.podcastsListFolderButtonTapped)
-
     }
 
-    private func oldFolderCreationFlow(from vc: UIViewController) {
+    func showUpsellIfNeeded(from vc: UIViewController) {
+        guard FeatureFlag.suggestedFolders.enabled,
+              !SubscriptionHelper.hasActiveSubscription(),
+              DateUtil.hasEnoughTimePassed(since: startingTime, time: Constants.intervalAfterStartup),
+              Settings.suggestedFoldersUpsellCount < Constants.maxUpsellDisplays,
+              DateUtil.hasEnoughTimePassed(since: Settings.suggestedFoldersLastUpsellDate, time: Constants.intervalBetweenUpsell),
+              dataManager.allPodcasts(includeUnsubscribed: false, reloadFromDatabase: false).count > Constants.minimumNumberOfPodcasts,
+              suggestedFoldersModel.loadingState == .loaded
+        else {
+            return
+        }
+        currentUpsellFlow = .cta
+        showUpsellSuggestedFolder(from: vc, fromUserAction: false, source: .suggestedFolderPopup)
+    }
+
+    private func manualFolderCreationFlow(from vc: UIViewController) {
         if !SubscriptionHelper.hasActiveSubscription() {
             navigationManager.showUpsellView(from: vc, source: .folders)
             return
         }
 
-        let creatFolderView = CreateFolderView { [weak vc] folderUuid in
-            if let folderUuid = folderUuid, let folder = DataManager.sharedManager.findFolder(uuid: folderUuid) {
+        let creatFolderView = CreateFolderView { [weak vc, weak self] folderUuid in
+            guard let self = self else { return }
+            if let folderUuid = folderUuid, let folder = dataManager.findFolder(uuid: folderUuid) {
                 vc?.dismiss(animated: true, completion: { [weak self] in
                     self?.navigationManager.navigateTo(NavigationManager.folderPageKey, data: [NavigationManager.folderKey: folder])
                 })
@@ -45,16 +87,13 @@ class FoldersCoordinator: NSObject {
         vc.present(hostingController, animated: true, completion: nil)
     }
 
-    private func newSuggestedFolderCreationFlow(from vc: UIViewController) {
+    private func suggestedFolderCreationFlow(from vc: UIViewController, source: AnalyticsSource) {
         if !SubscriptionHelper.hasActiveSubscription() {
-            if !SyncManager.isUserLoggedIn() {
-                navigationManager.showUpsellView(from: vc, source: .folders)
-            } else {
-                showUpSellSuggestedFolder(from: vc)
-            }
+            currentUpsellFlow = .userInitiated
+            showUpsellSuggestedFolder(from: vc, source: source)
             return
         }
-        let suggestedFoldersView = SuggestedFoldersView { [weak vc, weak self] result in
+        let suggestedFoldersView = SuggestedFoldersView(model: suggestedFoldersModel, source: source) { [weak vc, weak self] result in
             guard let self, let vc else { return }
 
             switch result {
@@ -78,35 +117,35 @@ class FoldersCoordinator: NSObject {
         hostingController.sheetPresentationController?.delegate = self
     }
 
-    private func showUpSellSuggestedFolder(from vc: UIViewController) {
-        let upsellSuggestedFoldersView = SuggestedFoldersUpSellView(model: SuggestedFoldersModel(failedToLoadAction: {[weak vc] in
-            guard let vc else { return }
-            vc.dismiss(animated: false) { [weak self] in
-                self?.navigationManager.showUpsellView(from: vc, source: .folders)
+    private func showUpsellSuggestedFolder(from vc: UIViewController, fromUserAction: Bool = false, source: AnalyticsSource) {
+        let suggestedFoldersView = SuggestedFoldersView(model: suggestedFoldersModel, source: source) { [weak vc, weak self] result in
+            guard let self, let vc else { return }
+            switch result {
+            case .dismiss:
+                vc.dismiss(animated: true)
+                //Update settings only if this was show by system
+                if !fromUserAction {
+                    Settings.suggestedFoldersLastUpsellDate = Date.now
+                    Settings.suggestedFoldersUpsellCount += 1
+                }
+                return
+            case .applySuggestedFolders, .createdManualFolder:
+                //Show upsell flow
+                startUpsellFlow(from: vc, source: source)
+                return
             }
-        }))
-        let hostingController = PCHostingController(rootView: upsellSuggestedFoldersView.environmentObject(Theme.sharedTheme))
-        if UIDevice.current.userInterfaceIdiom == .phone {
-            if #available(iOS 16.0, *) {
-                hostingController.sheetPresentationController?.detents = [.custom(resolver: { context in
-                    return context.maximumDetentValue * 0.65
-                })]
-            } else {
-                hostingController.sheetPresentationController?.detents = [.medium()]
-            }
-            hostingController.sheetPresentationController?.prefersGrabberVisible = true
-        } else {
-            hostingController.modalPresentationStyle = .formSheet
         }
+        let hostingController = PCHostingController(rootView: suggestedFoldersView.environmentObject(Theme.sharedTheme))
         vc.present(hostingController, animated: true, completion: nil)
+        hostingController.sheetPresentationController?.delegate = self
 
     }
 
     private func applySuggestedFolders(_ suggestedFolders: [SuggestedFolder]) {
-        DataManager.sharedManager.clearAllFolderInformation()
+        DataManager.sharedManager.deleteAllFoldersAndMarkSync()
         for suggestedFolder in suggestedFolders {
             let folder = makeFolder(from: suggestedFolder)
-            dataManager.bulkSetFolderUuid(folderUuid: folder.uuid, podcastUuids: suggestedFolder.topPodcastUuids)
+            dataManager.bulkSetFolderUuid(folderUuid: folder.uuid, podcastUuids: suggestedFolder.podcastUuids)
         }
         NotificationCenter.postOnMainThread(notification: ServerNotifications.podcastsRefreshed, object: nil)
     }
@@ -124,10 +163,67 @@ class FoldersCoordinator: NSObject {
         dataManager.save(folder: folder)
         return folder
     }
+
+    private func startUpsellFlow(from vc: UIViewController, source: AnalyticsSource) {
+        currentVC = vc
+        currentSource = source
+        addObservers()
+        vc.dismiss(animated: false) {
+            self.navigationManager.showUpsellView(from: vc, source: .folders, flow: .suggestedFolderUpsell)
+        }
+    }
+
+    private var cancellables = Set<AnyCancellable>()
+    private func addObservers() {
+        // Observe IAP flows notification
+        Publishers.Merge3(
+            NotificationCenter.default.publisher(for: ServerNotifications.iapPurchaseFailed),
+            NotificationCenter.default.publisher(for: ServerNotifications.iapPurchaseCancelled),
+            NotificationCenter.default.publisher(for: ServerNotifications.iapPurchaseCompleted)
+        )
+        .receive(on: OperationQueue.main)
+        .sink { [unowned self] notification in
+            refreshAfterUpsellFlow()
+        }
+        .store(in: &cancellables)
+
+        //Observe Login/Signup notification
+        NotificationCenter.default.publisher(for: .onboardingFlowDidDismiss)
+        .receive(on: OperationQueue.main)
+        .sink { [unowned self] notification in
+            refreshAfterUpsellFlow()
+        }
+        .store(in: &cancellables)
+    }
+
+    private func refreshAfterUpsellFlow() {
+        guard FeatureFlag.suggestedFolders.enabled,
+              SubscriptionHelper.hasActiveSubscription(),
+              let currentVC
+        else {
+            currentVC = nil
+            currentUpsellFlow = .none
+            cancellables = []
+            return
+        }
+        cancellables = []
+        currentUpsellFlow = .none
+        suggestedFolderCreationFlow(from: currentVC, source: currentSource)
+        self.currentVC = nil
+    }
 }
 
 extension FoldersCoordinator: UISheetPresentationControllerDelegate {
     func presentationControllerDidDismiss(_ presentationController: UIPresentationController) {
-        Analytics.track(.suggestedFoldersModalDismissed, properties: [:])
+        if currentUpsellFlow == .none {
+            Analytics.track(.suggestedFoldersPageDismissed, properties: [:])
+        } else {
+            if currentUpsellFlow == .cta {
+                Settings.suggestedFoldersLastUpsellDate = Date.now
+                Settings.suggestedFoldersUpsellCount += 1
+            }
+        }
+        currentUpsellFlow = .none
+        currentVC = nil
     }
 }