+++ podcasts/TranscriptShelfButton.swift
@@ -2,6 +2,7 @@ import UIKit
 import PocketCastsDataModel
 
 class TranscriptShelfButton: UIButton, CheckTranscriptAvailability {
+    var hasGeneratedTranscripts: Bool = false
     var isTranscriptEnabled: Bool {
         didSet {
             imageView?.tintColor = isTranscriptEnabled ? ThemeColor.playerContrast02() : ThemeColor.playerContrast06()
@@ -26,6 +27,7 @@ class TranscriptShelfButton: UIButton, CheckTranscriptAvailability {
 
 protocol CheckTranscriptAvailability: AnyObject {
     var isTranscriptEnabled: Bool { get set }
+    var hasGeneratedTranscripts: Bool { get set }
 
     func addTranscriptObservers()
     func checkTranscriptAvailability()
@@ -36,11 +38,13 @@ extension CheckTranscriptAvailability {
         NotificationCenter.default.addObserver(forName: Constants.Notifications.episodeTranscriptAvailabilityChanged, object: nil, queue: .main) { [weak self] notification in
             guard let episodeUuid = notification.userInfo?["episodeUuid"] as? String,
                   let isAvailable = notification.userInfo?["isAvailable"] as? Bool,
+                  let hasGeneratedTranscripts = notification.userInfo?["hasGeneratedTranscripts"] as? Bool,
                   episodeUuid == PlaybackManager.shared.currentEpisode()?.uuid else {
                 return
             }
 
             self?.isTranscriptEnabled = isAvailable
+            self?.hasGeneratedTranscripts = hasGeneratedTranscripts
         }
 
         NotificationCenter.default.addObserver(forName: Constants.Notifications.playbackTrackChanged, object: nil, queue: .main) { [weak self] notification in
@@ -50,6 +54,7 @@ extension CheckTranscriptAvailability {
 
     func checkTranscriptAvailability() {
         isTranscriptEnabled = false
+        hasGeneratedTranscripts = false
         let currentEpisode = PlaybackManager.shared.currentEpisode() as? Episode
         currentEpisode?.checkTranscriptAvailability()
     }